moved geoutils to script api
