+++ common/script/script-api/src/main/java/org/thingsboard/script/api/Perimeter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.geo;
+package org.thingsboard.script.api;
 
 import lombok.Data;
