+++ common/script/script-api/src/main/java/org/thingsboard/script/api/PerimeterType.java
@@ -13,12 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.geo;
+package org.thingsboard.script.api;
 
-import lombok.Data;
-
-@Data
-public class Coordinates {
-    private final double latitude;
-    private final double longitude;
+public enum PerimeterType {
+    CIRCLE, POLYGON
 }
