+++ common/script/script-api/src/main/java/org/thingsboard/script/api/RangeUnit.java
@@ -13,8 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.rule.engine.geo;
+package org.thingsboard.script.api;
 
-public enum PerimeterType {
-    CIRCLE, POLYGON
+public enum RangeUnit {
+    METER(1000.0), KILOMETER(1.0), FOOT(3280.84), MILE(0.62137), NAUTICAL_MILE(0.539957);
+
+    private final double fromKm;
+
+    RangeUnit(double fromKm) {
+        this.fromKm = fromKm;
+    }
+
+    public double fromKm(double v) {
+        return v * fromKm;
+    }
 }
